cleaning logic for impersonations

+++ src/main/java/com/web/back/model/responses/ImpersonateResponse.java
@@ -1,4 +1,4 @@
 package com.web.back.model.responses;
 
-public record ImpersonateResponse(Integer impersonationId, String actorUserName, String targetUserName){
+public record ImpersonateResponse(Long impersonationId, String actorUserName, String targetUserName){
 }
